requirements.txt
langgraph
langchain
langchain-openai
langchain-community
redis
pymupdf
python-dotenv
openai
pydantic
psycopg2-binary
Pillow
langchain-text-splitters
